Working towards enrichment charts working for all enrichment tables and not only for one at a time.
Added an advanced options group of tunables and cleaned up the user
dialog options. Also checking for the correct cluster column now. 
Fixed the right-click menu items in the enrichment tables using
anonymous action listener classes. 
Initiated the showEnrichment task factory for the cluster enrichment to
fix the not-showing of the panel.